refact. some changes and code upgrades

+++ app/src/main/java/com/puroblast/weatherappcompose/network/model/Wind.kt
@@ -2,5 +2,5 @@ package com.puroblast.weatherappcompose.network.model
 
 data class Wind(
     val deg: Int,
-    val speed: Int
+    val speed: Double
 )
